Move code into Shakespeare, rm unused
