+++ backend/shakespeare/analyze.py
@@ -23,7 +23,17 @@ def build_shakespeare_model(lines, model_name):
     write_model(model, model_name)
     return model
 
+def build_combined_shakespeare_model(models):
+    combined_model = None
+    for model in models:
+        if not combined_model:
+            combined_model = model
+        else:
+            combined_model = combine_models([combined_model, model])
+    write_model(combined_model, name=shakespeare_model_name("combined"))
 
+def shakespeare_model_name(archetype):
+    return "#{}".format(archetype)
 
 def load_shakespeare_model(archetype):
-    return load_model("#" + archetype)
+    return load_model(shakespeare_model_name(archetype))
